requirements.txt
Pillow==10.3.0
requests==2.31.0
python-dotenv==1.0.1
urllib3==2.2.1
instagrapi==2.1.2
